+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next release
 
+- Vertically-center `Option` contents
 - Add new real-world example – approval poll from 538
   (([#140](https://github.com/graph-paper-org/graph-paper/pull/140)))
 - Add `AxisContainer` to exports in `@graph-paper/guide`